Project structures refactoring
Some folders and components are renamed or moved to another folder in order to make it easy to read
