+++ src/views/AddCard/AddCard.js
@@ -9,7 +9,7 @@ const AddCard = () => {
             </Text>
         </View>
     )
-}
+};
 
 export default AddCard
